Updated schema and resolvers for pagination

+++ src/resolvers/Mutation/post.ts
@@ -1,21 +1,46 @@
 import { Context } from 'src/types';
+// import { Post } from '@prisma/client';
+
+interface PostCreateArgs {
+	title: string;
+	text: string;
+}
+
+interface PostUpdateArgs {
+	id: string;
+	title?: string;
+	text?: string;
+}
+
+// interface PostPayloadType {
+// 	errors: {
+// 		field: string;
+// 		message: string;
+// 	}[];
+// 	post: Post | null;
+// }
 
 export const postMutationResolvers = {
-	postCreate: async (_parent: any, { title, text }: { title: string; text: string }, { prisma }: Context) => {
+	postCreate: async (_parent: any, { title, text }: PostCreateArgs, { prisma, req }: Context) => {
+		// user is not logged in
+		if (!req.session.userId) {
+			throw new Error('Not Authenticated');
+		}
+
 		const post = await prisma.post.create({
 			data: {
 				title,
 				text,
-				creatorId: 1, // Mocking the logged in user
+				creatorId: req.session.userId, // Only logged-in user can post
 			},
 		});
 		return post;
 	},
-	postUpdate: async (
-		_parent: any,
-		{ id, title, text }: { id: string; title: string | undefined; text: string | undefined },
-		{ prisma }: Context
-	) => {
+	postUpdate: async (_parent: any, { id, title, text }: PostUpdateArgs, { prisma, req }: Context) => {
+		if (!req.session.userId) {
+			throw new Error('Not Authenticated');
+		}
+
 		const postExists = await prisma.post.findUnique({
 			where: {
 				id: Number(id),
@@ -24,6 +49,10 @@ export const postMutationResolvers = {
 
 		if (!postExists) {
 			return null;
+		}
+
+		if (postExists.creatorId !== req.session.userId) {
+			throw new Error('Unauthorized');
 		}
 
 		const updatedPost = await prisma.post.update({
@@ -35,9 +64,29 @@ export const postMutationResolvers = {
 				text,
 			},
 		});
+
 		return updatedPost;
 	},
-	postDelete: async (_parent: any, { id }: { id: string }, { prisma }: Context) => {
+	postDelete: async (_parent: any, { id }: { id: string }, { prisma, req }: Context) => {
+		if (!req.session.userId) {
+			throw new Error('Not Authenticated');
+		}
+
+		const postExists = await prisma.post.findUnique({
+			where: {
+				id: Number(id),
+			},
+		});
+
+		// REFACTOR
+		if (!postExists) {
+			return false;
+		}
+
+		if (postExists.creatorId !== req.session.userId) {
+			throw new Error('Unauthorized');
+		}
+
 		try {
 			await prisma.post.delete({
 				where: {
